4.221

+++ Class/Json/JsonLocalPlayer.cs
@@ -2,8 +2,8 @@
 public class JsonLocalPlayer : Json
 {
     public string map_id { get; set; }
-    public int scrap { get; set; }
-    public int credits { get; set; }
+    public double scrap { get; set; }
+    public double credits { get; set; }
     public int experience { get; set; }
     public int level { get; set; }
     public int ranking_points { get; set; }
